+++ dags/veda_data_pipeline/veda_process_raster_pipeline.py
@@ -19,34 +19,31 @@ This DAG is supposed to be triggered by `veda_discover`. But you still can trigg
     "filename_regex": "^(.*).tif$",
     "discovery": "s3",
     "datetime_range": "month",
-    "upload": false,
-    "cogify": false,
     "discovered": 33,
-    "payload": "s3://veda-uah-sit-mwaa-853558080719/events/geoglam/s3_discover_output_6c46b57a-7474-41fe-977a-19d164531cdc.json"
+    "payload": "s3://veda-uah-sit-mwaa-853558080719/events/geoglam/s3_discover_output_6c46b57a-7474-41fe-977a-.json"
 }	
 ```
 - [Supports linking to external content](https://github.com/NASA-IMPACT/veda-data-pipelines)
 """
 
-templat_dag_run_conf = {
+template_dag_run_conf = {
     "collection": "<collection_name>",
     "prefix": "<prefix>/",
     "bucket": "<bucket>",
     "filename_regex": "<filename_regex>",
-    "discovery": "<s3>|cmr",
+    "discovery": "<s3>",
     "datetime_range": "<month>|<day>",
-    "upload": "<false> | true",
-    "cogify": "false | true",
     "payload": "<s3_uri_event_payload",
 }
 dag_args = {
+    "max_active_runs": 20,
     "start_date": pendulum.today("UTC").add(days=-1),
     "schedule_interval": None,
     "catchup": False,
     "doc_md": dag_doc_md,
 }
 
-with DAG(dag_id="veda_ingest_raster", params=templat_dag_run_conf, **dag_args) as dag:
+with DAG(dag_id="veda_ingest_raster", params=template_dag_run_conf, **dag_args) as dag:
     start = DummyOperator(task_id="Start", dag=dag)
     end = DummyOperator(task_id="End", trigger_rule=TriggerRule.ONE_SUCCESS, dag=dag)