+++ infrastructure/outputs.tf
@@ -7,3 +7,9 @@ output "mwaa_s3_name" {
 output "mwaa_subnets" {
   value = module.mwaa.subnets
 }
+output "airflow_env" {
+  value = module.mwaa.mwaa_environment_name
+}
+output "workflows_api" {
+  value = aws_apigatewayv2_api.workflows_http_api.api_endpoint
+}